// src/enums/codes.ts
export enum CODES {
  UNAUTHORIZED = "UNAUTHORIZED",
  COURSE_DOES_NOT_EXIST = "COURSE_DOES_NOT_EXIST",
  BAD_REQUEST = "BAD_REQUEST",
  UNEXPECTED_ERROR = "UNEXPECTED_ERROR",
  NOT_FOUND = "NOT_FOUND",
  TOKEN_EXPIRED = "TOKEN_EXPIRED",
  VALIDATION_REQUEST_ERROR = "VALIDATION_REQUEST_ERROR",
  EMAIL_IN_USE = "EMAIL_IN_USE",
  ACCESS_TOKEN_EXPIRED = "ACCESS_TOKEN_EXPIRED",
  REFRESH_TOKEN_EXPIRED = "REFRESH_TOKEN_EXPIRED",
  REUSE_DETECTION = "REUSE_DETECTION",
  NO_ACCESS_TOKEN = "NO_ACCESS_TOKEN",
  NO_REFRESH_TOKEN = "NO_REFRESH_TOKEN",
  UNABLE_TO_LOGIN = "UNABLE_TO_LOGIN",
  MULTER_IMAGE_TYPE_ERROR = "MULTER_IMAGE_TYPE_ERROR",
  MULTER_SIZE_ERROR = "MULTER_SIZE_ERROR",
  MULTER_FILE_DOES_NOT_EXIST = "MULTER_FILE_DOES_NOT_EXIST",
  MULTER_TOO_MANY_IMAGES = "MULTER_TOO_MANY_IMAGES",
  VIDEO_CONVERTED_ALREADY = "VIDEO_CONVERTED_ALREADY",
  VIDEO_IS_STILL_BEING_CONVERTED = "VIDEO_IS_STILL_BEING_CONVERTED",
  NEW_PASSWORD_DOES_NOT_MATCH_OLD_PASSWORD = "NEW_PASSWORD_DOES_NOT_MATCH_OLD_PASSWORD",
  PASSWORDS_MUST_BE_THE_SAME = "PASSWORDS_MUST_BE_THE_SAME",
  ACCOUNT_NOT_ACTIVATED = "ACCOUNT_NOT_ACTIVATED",
  EXPIRED = "EXPIRED",
  SUCCESS = "SUCCESS",
  ACCOUNT_DELETED = "ACCOUNT_DELETED",
  ADMIN_ONLY = "ADMIN_ONLY",
  NON_SUPER_USERS_ONLY = "NON_SUPER_USERS_ONLY",
  CANNOT_DECODE_TOKEN = "CANNOT_DECODE_TOKEN",
  NOT_ALLOWED = "NOT_ALLOWED",
  ROUTE_DOES_NOT_EXIST = "ROUTE_DOES_NOT_EXIST",
  MAX_USERS_IN_SYSTEM = "MAX_USERS_IN_SYSTEM",
  NO_ID = "NO_ID",
  GOOGLE_AUTH_ERROR = "GOOGLE_AUTH_ERROR",
  PASSWORD_EXIST_ALREADY = "PASSWORD_EXIST_ALREADY",
  NO_PASSWORD_TO_ACCOUNT = "NO_PASSWORD_TO_ACCOUNT",
  SCHOOL_EXIST_ALREADY = "SCHOOL_EXIST_ALREADY",
  IMAGE_UPLOAD_ERROR = "IMAGE_UPLOAD_ERROR",
  PICTURE_DOES_NOT_EXIST = "PICTURE_DOES_NOT_EXIST",
  CANNOT_DELETE_ALL_IMAGES = "CANNOT_DELETE_ALL_IMAGES",
  DUPLICATE_RATING = "DUPLICATE_RATING",
  NOT_YOUR_RATING = "NOT_YOUR_RATING",
  SCHOOL_DOES_NOT_EXIST = "SCHOOL_DOES_NOT_EXIST",
  PROGRAM_DOES_NOT_EXIST = "PROGRAM_DOES_NOT_EXIST",
  SCHOOL_PROGRAM_EXISTS_ALREADY = "SCHOOL_PROGRAM_EXISTS_ALREADY",
  PROGRAM_COURSE_EXISTS_ALREADY = "PROGRAM_COURSE_EXISTS_ALREADY",
  PROGRAM_COURSE_DOES_NOT_EXIST = "PROGRAM_COURSE_DOES_NOT_EXIST",
  SCHOOL_PROGRAM_DOES_NOT_EXIST = "SCHOOL_PROGRAM_DOES_NOT_EXIST",
  TOKEN_DOES_NOT_EXIST = "TOKEN_DOES_NOT_EXIST",
  TWO_FACTOR_ENABLED_ALREADY = "TWO_FACTOR_ENABLED_ALREADY",
  TWO_FACTOR_NOT_YET_ENABLED = "TWO_FACTOR_NOT_YET_ENABLED",
  WRONG_CODE_OR_EXPIRED_CODE = "WRONG_CODE_OR_EXPIRED_CODE",
  NOT_ALLOWED_USER_TYPE = "NOT_ALLOWED_USER_TYPE",
  TOKEN_DOES_NOT_MATCH = "TOKEN_DOES_NOT_MATCH",
  CANNOT_DELETE_YOURSELF = "CANNOT_DELETE_YOURSELF",
  CANNOT_DEACTIVATE_YOURSELF = "CANNOT_DEACTIVATE_YOURSELF",
}